feat: compact

+++ file/file.go
@@ -4,7 +4,7 @@ import "io"
 
 // Options
 type Options struct {
-	FID      uint32
+	FID      uint64
 	FileName string
 	Dir      string
 	Path     string
